Moved parsers.hpp to the serialization directory
