+++ source/cell/cell_reference.cpp
@@ -30,7 +30,7 @@
 #include <xlnt/worksheet/range_reference.hpp>
 
 #include <detail/constants.hpp>
-#include <detail/parsers.hpp>
+#include <detail/serialization/parsers.hpp>
 
 namespace xlnt {
