+++ tests/internal/parser_test_suite.cpp
@@ -21,7 +21,7 @@
 // @license: http://www.opensource.org/licenses/mit-license.php
 // @author: see AUTHORS file
 
-#include <detail/parsers.hpp>
+#include <detail/serialization/parsers.hpp>
 #include "helpers/assertions.hpp"
 
 #include <helpers/test_suite.hpp>
